+++ CHANGELOG.md
@@ -1,14 +1,19 @@
 # Changelog
 
 ## Unreleased
+### Added
+- Add sample rate selection in wav export
+
 ### Changed
 - [#152] - Use RtAudio for sound processing ([#96]; thanks [@jpcima], [@OPNA2608] and others)
 
 ### Fixed
 - Fix to draw the playing cursor when follow mode is turned off
 - [#152] - Prevent real step size setting to 0 (thanks [@OPNA2608])
+- [#153] - Fix hardware envelope to set data in sequence (thanks [@papiezak])
 
 [#152]: https://github.com/rerrahkr/BambooTracker/pull/152
+[#153]: https://github.com/rerrahkr/BambooTracker/issues/153
 
 ## v0.2.4 (2019-09-17)
 ### Changed